requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
google-generativeai==0.3.2